recipes/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
djangorestframework==3.15.2
sqlparse==0.5.3
typing-extensions==4.13.2
django-cors-headers==4.3.1
